Extract download methods to DownloadService
